Clone submodules in dev container after creation (#8)
So users don't have to do this manually, when using the container.

If one attempts to configure or build the project with CMake
(including through VS Code when it offeres to configure
automatically) without doing this first, the error message is very
unintuitive: cmake does not find ninja.

This uses postStartCommand to clone only once per container creation,
and uses waitFor so wait to enter the container until this has been
completed. Without waitFor, the clone runs asynchronously and the
vcpkg subdirectory is confusingly unpopulated for a while after the
container is entered.