Build project in dev container onCreate
Among other benefits, this will facilitate prebuilds that are
immediately ready to use. However, it does greatly increase the
amount of time the container takes to create when not using
prebuilds -- although, in many uses, one would then take that same
amount of time, with greater annoyance, to manually trigger the
build before doing much else.

The real reason to do this is to *configure* the project. Building
causes that to happen, and is less confusing than only configuring
(because one may wonder what was happening, if there is no usable
binary). Configuring resolves, downloads, **and builds** vcpkg
dependencies in the submodule, which is why it takes so long and is
an area where prebuilds will be very helpful.